requirements.txt
flask>=2.3.3
flask-restx>=0.5.1
flask-limiter>=3.5.0
flask-cors>=4.0.0
pillow>=10.0.0
tensorflow>=2.14.0
numpy>=1.24.0
python-dotenv>=1.0.0
pytest>=7.4.0
opencv-python-headless==4.8.0.74  # Using headless version for server deployment
mtcnn==0.1.1
gunicorn==21.2.0
